update all imports from moving firestore related functions out of components folder
